feat(e2e-tests) simplify e2e test structure
