+++ packages/core/package.json
@@ -25,7 +25,7 @@
   "scripts": {
     "build": "stencil build",
     "dev": "sd concurrent \"stencil build --dev --watch\" \"stencil-dev-server\"",
-    "e2e": "node ./scripts/e2e.js",
+    "e2e": "node ./scripts/run-e2e.js",
     "snapshot": "node ./scripts/e2e.js --snapshot",
     "test": "jest --no-cache",
     "test.watch": "jest --watch --no-cache",
