+++ packages/core/scripts/run-e2e.js
@@ -1,0 +1,3 @@
+'use strict';
+
+require('./e2e-test-runner').run();
